CS-2479 - Cancel delete and non-deletable fix
Modals don't seem to function correctly as far as scroll bars are
concerned when placed in a sub-component from where they are activated
